refactor: extract template library components
